requirements.txt
certifi==2019.11.28
chardet==3.0.4
dateutils==0.6.7
Django==2.2
django-rest-framework==0.1.0
django-rest-framework-braces==0.3.4
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
get==2019.4.13
idna==2.8
numpy==1.17.4
pandas==0.25.3
post==2019.4.13
pprint==0.1
public==2019.4.13
python-dateutil==2.8.1
python-stdnum==1.12
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
